+++ lib/db_sucker/application/sklaven_treiber/worker.rb
@@ -18,17 +18,17 @@ module DbSucker
           @monitor = Monitor.new
           @timings = {}
           @perform = %w[].tap do |perform|
-            perform << "r_dump_file"
-            perform << "r_calculate_raw_hash" if ctn.integrity?
-            perform << "r_compress_file"
-            perform << "r_calculate_compressed_hash" if ctn.integrity?
-            perform << "l_download_file"
-            perform << "l_verify_compressed_hash" if ctn.integrity?
-            perform << "l_copy_file" if var.copies_file? && var.copies_file_compressed?
-            if var.requires_uncompression?
-              perform << "l_decompress_file"
-              perform << "l_verify_raw_hash" if ctn.integrity?
-              perform << "l_copy_file" if var.copies_file? && !var.copies_file_compressed?
+            perform << "r_dump_file"                                                                  # PV status (bytes + seconds)
+            perform << "r_calculate_raw_hash" if ctn.integrity?                                       # PV status
+            perform << "r_compress_file"                                                              # --------- DONE! (has pv + sec-fallback)
+            perform << "r_calculate_compressed_hash" if ctn.integrity?                                # PV status
+            perform << "l_download_file"                                                              # --------- DONE! (always full status)
+            perform << "l_verify_compressed_hash" if ctn.integrity?                                   # PV status? we would need local PV or we lockin on SHA and do it with ruby locally
+            perform << "l_copy_file" if var.copies_file? && var.copies_file_compressed?               # --------- DONE! (always full status)
+            if var.requires_uncompression?                                                            # ---------
+              perform << "l_decompress_file"                                                          # --------- DONE! (always full status)
+              perform << "l_verify_raw_hash" if ctn.integrity?                                        # PV status? we would need local PV or we lockin on SHA and do it with ruby locally
+              perform << "l_copy_file" if var.copies_file? && !var.copies_file_compressed?            # --------- DONE! (always full status)
               perform << "l_import_file" if var.data["database"]
             end
           end