+++ lib/db_sucker/application/window.rb
@@ -6,16 +6,17 @@ module DbSucker
       COL1 = 20
       COL2 = 25
       COL3 = 20
+      OutputHelper.hook(self)
+
       attr_reader :app, :sklaventreiber
       attr_accessor :view
-      OutputHelper.hook(self)
 
       def initialize app, sklaventreiber
         @app = app
         @sklaventreiber = sklaventreiber
         @monitor = Monitor.new
-        @l = 0
-        @t = 0
+        @l = 0 # line
+        @t = 0 # tick
         @view = :status
       end
 
@@ -37,6 +38,19 @@ module DbSucker
 
       def line l = 1
         setpos(l - 1, 0)
+      end
+
+      [:red, :blue, :yellow, :cyan, :magenta, :gray, :green, :white].each do |c|
+        define_method(c) do |*args, &block|
+          color = self.class.const_get "COLOR_#{c.to_s.upcase}"
+          attron(color_pair(color)|A_NORMAL) do
+            if block
+              block.call
+            else
+              args.each {|a| addstr(a) }
+            end
+          end
+        end
       end
 
       def next_line
@@ -62,12 +76,12 @@ module DbSucker
 
         update do
           next_line
-          attron(color_pair(COLOR_RED)|A_NORMAL) { addstr("RenderError occured!") }
+          red "RenderError occured!"
           next_line
-          attron(color_pair(COLOR_RED)|A_NORMAL) { addstr("#{ex.class}: #{ex.message}") }
+          red "#{ex.class}: #{ex.message}"
           ex.backtrace.each do |l|
             next_line
-            attron(color_pair(COLOR_RED)|A_NORMAL) { addstr("    #{l}") }
+            red("    #{l}")
           end
         end
         sleep 1
@@ -113,47 +127,44 @@ module DbSucker
 
       def _view_status
         next_line
-        attron(color_pair(COLOR_YELLOW)|A_NORMAL) { addstr("        Status: ") }
-        attron(color_pair(self.class.const_get "COLOR_#{sklaventreiber.status[1].to_s.upcase.presence || "BLUE"}")|A_NORMAL) { addstr(sklaventreiber.status[0]) }
+        yellow "        Status: "
+        send(sklaventreiber.status[1].presence || :blue, sklaventreiber.status[0])
         # shutdown message
         if $core_runtime_exiting && sklaventreiber.status[0] != "terminated"
-          attron(color_pair(COLOR_RED)|A_BOLD) { addstr(" (HALTING … please wait)") }
+          red " (HALTING … please wait)"
         end
 
         next_line
-        attron(color_pair(COLOR_YELLOW)|A_NORMAL) { addstr("       Threads: ") }
-        attron(color_pair(COLOR_BLUE)|A_NORMAL) { addstr("#{Thread.list.length} ".ljust(COL1, " ")) }
-
-        # attron(color_pair(COLOR_YELLOW)|A_NORMAL) { addstr("Connections: ") }
-        # attron(color_pair(COLOR_BLUE)|A_NORMAL) { addstr("#{Thread.list.length} ".ljust(COL1, " ")) }
+        yellow "       Threads: "
+        blue "#{Thread.list.length} ".ljust(COL1, " ")
 
         next_line
-        attron(color_pair(COLOR_YELLOW)|A_NORMAL) { addstr("       Started: ") }
-        attron(color_pair(COLOR_BLUE)|A_NORMAL) { addstr("#{@app.boot}") }
-        attron(color_pair(COLOR_YELLOW)|A_NORMAL) { addstr(" (") }
-        attron(color_pair(COLOR_BLUE)|A_NORMAL) { addstr(human_seconds(Time.current - app.boot)) }
-        attron(color_pair(COLOR_YELLOW)|A_NORMAL) { addstr(")") }
+        yellow "       Started: "
+        blue "#{@app.boot}"
+        yellow " ("
+        blue human_seconds(Time.current - app.boot)
+        yellow ")"
 
         next_line
-        attron(color_pair(COLOR_YELLOW)|A_NORMAL) { addstr("Transaction ID: ") }
-        attron(color_pair(COLOR_CYAN)|A_NORMAL) { addstr(sklaventreiber.trxid) }
+        yellow "Transaction ID: "
+        cyan sklaventreiber.trxid
 
         next_line
-        attron(color_pair(COLOR_YELLOW)|A_NORMAL) { addstr("      Database: ") }
-        attron(color_pair(COLOR_MAGENTA)|A_NORMAL) { addstr(sklaventreiber.data[:database] || "?") }
-        attron(color_pair(COLOR_YELLOW)|A_NORMAL) { addstr(" (transfering ") }
-        attron(color_pair(COLOR_BLUE)|A_NORMAL) { addstr("#{sklaventreiber.data[:tables_transfer] || "?"}") }
-        attron(color_pair(COLOR_YELLOW)|A_NORMAL) { addstr(" of ") }
-        attron(color_pair(COLOR_BLUE)|A_NORMAL) { addstr("#{sklaventreiber.data[:tables_total] || "?"}") }
-        attron(color_pair(COLOR_YELLOW)|A_NORMAL) { addstr(" tables)") }
+        yellow "      Database: "
+        magenta sklaventreiber.data[:database] || "?"
+        yellow " (transfering "
+        blue "#{sklaventreiber.data[:tables_transfer] || "?"}"
+        yellow " of "
+        blue "#{sklaventreiber.data[:tables_total] || "?"}"
+        yellow " tables)"
 
         next_line
         total, done = sklaventreiber.data[:tables_transfer], sklaventreiber.data[:tables_done]
         perc = total && done ? f_percentage(done, total) : "?"
-        attron(color_pair(COLOR_YELLOW)|A_NORMAL) { addstr("      Progress: ") }
-        attron(color_pair(COLOR_GREEN)|A_NORMAL) { addstr(perc) }
-        attron(color_pair(COLOR_YELLOW)|A_NORMAL) { addstr(" – ") }
-        attron(color_pair(COLOR_BLUE)|A_NORMAL) { addstr("#{done}/#{total || "?"} workers done") }
+        yellow "      Progress: "
+        green perc
+        yellow " – "
+        blue "#{done}/#{total || "?"} workers done"
 
         _render_workers
       end
@@ -171,7 +182,7 @@ module DbSucker
               part = rest.group_by(&:state).map do |k, v|
                 "#{v.length} #{k}"
               end
-              attron(color_pair(COLOR_YELLOW)|A_NORMAL) { addstr("… #{rest.length} more [#{part.join(", ")}]") }
+              yellow "… #{rest.length} more [#{part.join(", ")}]"
               break
             end
             _render_worker_line(w)
@@ -186,11 +197,11 @@ module DbSucker
 
         # status icon
         case worker.state
-          when :pending then attron(color_pair(COLOR_GRAY)|A_NORMAL) { addstr("⊙") }
-          when :aquired then attron(color_pair(COLOR_WHITE)|A_NORMAL) { addstr("⊙") }
-          when :done then attron(color_pair(COLOR_GREEN)|A_NORMAL) { addstr("✔") }
-          when :failed then attron(color_pair(COLOR_RED)|A_NORMAL) { addstr("✘") }
-          when :canceled then attron(color_pair(COLOR_RED)|A_NORMAL) { addstr("⊘") }
+          when :pending then gray("⊙")
+          when :aquired then white("⊙")
+          when :done then green("✔")
+          when :failed then red("✘")
+          when :canceled then red("⊘")
           when :running then
             c = case @t % 4
               when 0 then "◜" # "╭"
@@ -198,23 +209,21 @@ module DbSucker
               when 2 then "◞" # "╯"
               when 3 then "◟" # "╰"
             end
-            attron(color_pair(COLOR_YELLOW)|A_NORMAL) { addstr("#{c}") }
+            yellow "#{c}"
         end
 
         # table_name
-        attron(color_pair(worker.should_cancel ? COLOR_RED : COLOR_MAGENTA)|A_NORMAL) { addstr(" #{worker.table}".ljust(col1 + 1, " ")) }
-        attron(color_pair(COLOR_GRAY)|A_NORMAL) { addstr(" | ") }
+        send(worker.should_cancel ? :red : :magenta, " #{worker.table}".ljust(col1 + 1, " "))
+        gray " | "
 
         # status
         if worker.step
-          attron(color_pair(COLOR_CYAN)|A_NORMAL) { addstr("[#{worker.step}/#{worker.perform.length}] ") }
+          cyan "[#{worker.step}/#{worker.perform.length}] "
         end
-        attron(color_pair(self.class.const_get "COLOR_#{worker.status[1].to_s.upcase.presence || "BLUE"}")|A_NORMAL) do
-          if worker.status[0].respond_to?(:to_curses)
-            worker.status[0].to_curses(self)
-          else
-            addstr(uncolorize "#{worker.status[0]}")
-          end
+        if worker.status[0].respond_to?(:to_curses)
+          worker.status[0].to_curses(self)
+        else
+          send(worker.status[1].presence || :blue, uncolorize("#{worker.status[0]}"))
         end
       end